refactor connnection class
