+++ src/state/CallViewModel.ts
@@ -6,7 +6,6 @@ Please see LICENSE in the repository root for full details.
 */
 
 import {
-  connectedParticipantsObserver,
   observeParticipantEvents,
   observeParticipantMedia,
 } from "@livekit/components-core";
@@ -21,7 +20,6 @@ import {
 import E2EEWorker from "livekit-client/e2ee-worker?worker";
 import {
   ClientEvent,
-  type MatrixClient,
   RoomStateEvent,
   SyncState,
   type Room as MatrixRoom,
@@ -105,7 +103,6 @@ import { shallowEquals } from "../utils/array";
 import { calculateDisplayName, shouldDisambiguate } from "../utils/displayname";
 import { type MediaDevices } from "./MediaDevices";
 import { type Behavior } from "./Behavior";
-import { getSFUConfigWithOpenID } from "../livekit/openIDSFU";
 import { defaultLiveKitOptions } from "../livekit/options";
 import {
   enterRTCSession,
@@ -114,6 +111,7 @@ import {
 } from "../rtcSessionHelpers";
 import { E2eeType } from "../e2ee/e2eeType";
 import { MatrixKeyProvider } from "../e2ee/matrixKeyProvider";
+import { Connection, PublishConnection } from "./Connection";
 
 export interface CallViewModelOptions {
   encryptionSystem: EncryptionSystem;
@@ -436,88 +434,6 @@ function getRoomMemberFromRtcMember(
   return { id, member };
 }
 
-// TODO-MULTI-SFU Add all device syncing logic from useLivekit
-class Connection {
-  private readonly sfuConfig = getSFUConfigWithOpenID(
-    this.client,
-    this.serviceUrl,
-    this.livekitAlias,
-  );
-
-  public async startSubscribing(): Promise<void> {
-    this.stopped = false;
-    const { url, jwt } = await this.sfuConfig;
-    if (!this.stopped) await this.livekitRoom.connect(url, jwt);
-    await this.livekitRoom.localParticipant.publishTrack(tracks[0]);
-  }
-
-  public async startPublishing(): Promise<void> {
-    this.stopped = false;
-    const { url, jwt } = await this.sfuConfig;
-    if (!this.stopped) await this.livekitRoom.connect(url, jwt);
-
-    if (!this.stopped) {
-      const tracks = await this.livekitRoom.localParticipant.createTracks({
-        audio: true,
-        video: true,
-      });
-      for (const track of tracks) {
-        await this.livekitRoom.localParticipant.publishTrack(track);
-      }
-    }
-  }
-
-  private stopped = false;
-
-  public stop(): void {
-    void this.livekitRoom.disconnect();
-    this.stopped = true;
-  }
-
-  public readonly participantsIncludingSubscribers$ = this.scope.behavior(
-    connectedParticipantsObserver(this.livekitRoom),
-    [],
-  );
-
-  public readonly publishingParticipants$ = (
-    memberships$: Behavior<CallMembership[]>,
-  ): Observable<RemoteParticipant[]> =>
-    this.scope.behavior(
-      combineLatest([
-        connectedParticipantsObserver(this.livekitRoom),
-        memberships$,
-      ]).pipe(
-        map(([participants, memberships]) => {
-          const publishingMembers = membershipsFocusUrl(
-            memberships,
-            this.matrixRTCSession,
-          )
-            .filter((f) => f.livekit_service_url === this.serviceUrl)
-            .map((f) => f.membership);
-
-          const publishingP = publishingMembers
-            .map((m) => {
-              return participants.find((p) => {
-                return p.identity === `${m.sender}:${m.deviceId}`;
-              });
-            })
-            .filter((p): p is RemoteParticipant => !!p);
-          return publishingP;
-        }),
-      ),
-      [],
-    );
-
-  public constructor(
-    private readonly livekitRoom: LivekitRoom,
-    private readonly serviceUrl: string,
-    private readonly livekitAlias: string,
-    private readonly client: MatrixClient,
-    private readonly scope: ObservableScope,
-    private readonly matrixRTCSession: MatrixRTCSession,
-  ) {}
-}
-
 export class CallViewModel extends ViewModel {
   private readonly e2eeOptions = getE2eeOptions(
     this.options.encryptionSystem,
@@ -535,7 +451,7 @@ export class CallViewModel extends ViewModel {
 
   private readonly localConnection = this.localFocus.then(
     (focus) =>
-      new Connection(
+      new PublishConnection(
         this.localConnectionLivekitRoom,
         focus.livekit_service_url,
         this.livekitAlias,
@@ -1765,7 +1681,7 @@ export class CallViewModel extends ViewModel {
       .subscribe(
         (c) =>
           void c
-            .startPublishing()
+            .start()
             // eslint-disable-next-line no-console
             .then(() => console.log("successfully started publishing"))
             // eslint-disable-next-line no-console
@@ -1774,7 +1690,7 @@ export class CallViewModel extends ViewModel {
     this.connectionInstructions$
       .pipe(this.scope.bind())
       .subscribe(({ start, stop }) => {
-        for (const connection of start) void connection.startSubscribing();
+        for (const connection of start) void connection.start();
         for (const connection of stop) connection.stop();
       });
     combineLatest([this.localFocus, this.joined$])
@@ -1838,7 +1754,7 @@ export class CallViewModel extends ViewModel {
   }
 }
 
-const membershipsFocusUrl = (
+export const membershipsFocusUrl = (
   memberships: CallMembership[],
   matrixRTCSession: MatrixRTCSession,
 ): { livekit_service_url: string; membership: CallMembership }[] => {
